Made it so charges need to be authenticated.
Changed the create method in charges so that the charge does not
get returned to the client after it is created.  Also created a
script to test charges via the terminal and curl requests.

Curl requests cannot successfully create a token, but the checkout
widget on the client app successfully creates a token.